@@ -4,18 +4,29 @@
  * @author Khoa Bui (khoaofgod)  <khoaofgod@gmail.com> http://www.phpfastcache.com
  * @author Georges.L (Geolim4)  <contact@geolim4.com>
  */
-use phpFastCache\CacheManager;
+
+use phpFastCache\Helper\TestHelper;
+
 chdir(__DIR__);
-require_once '../src/autoload.php';
+require_once __DIR__ . '/../src/autoload.php';
+$testHelper = new TestHelper('Autoload');
 
-$status = 0;
-echo "Testing autoload\n";
-
-if(!class_exists('phpFastCache\CacheManager'))
-{
-    echo "[FAIL] Autoload failed to find the CacheManager\n";
-    $status = 1;
+/**
+ * Testing PhpFastCache autoload
+ */
+if (!class_exists('phpFastCache\CacheManager')) {
+    $testHelper->printFailText('Autoload failed to find the CacheManager');
+}else{
+    $testHelper->printPassText('Autoload successfully found the CacheManager');
 }
-echo "[PASS] Autoload successfully found the CacheManager\n";
 
-exit($status);
+/**
+ * Testing Psr autoload
+ */
+if (!interface_exists('Psr\Cache\CacheItemInterface')) {
+    $testHelper->printFailText('Autoload failed to find the Psr CacheItemInterface');
+}else{
+    $testHelper->printPassText('Autoload successfully found the Psr CacheItemInterface');
+}
+
+$testHelper->terminateTest();
